refactor: Make abstract dialog to reuse JSON common functions (#17)
